examples/spyglass: rewrite as two-runtime pattern with fifo
The keyboard notifier fires in hardirq context, which is incompatible with
the process-context runtime used by the device. Split into two scripts
following the systrack pattern:

- examples/spyglass/notifier.lua: hardirq runtime, registers the keyboard
  notifier and pushes each captured char into a shared fifo.
- examples/spyglass/device.lua: process runtime, creates the fifo, exposes
  /dev/spyglass, drains the fifo on read.

fifo fits the producer-consumer shape better than rcu.table: push is O(1)
per char (no string concatenation), the buffer is bounded (4 KB), and the
pop drains atomically so there is no read/clear race. If the consumer
falls behind and the fifo fills, the producer silently drops via pcall.

Drop the legacy UDP-forwarding mode: it relied on coupling the UDP send to
the read syscall, which made "configure and nothing happens until you cat"
an awkward surprise. The simpler device-only path is a better match for a
didactic example; users who need networking can compose it separately.

Update README and Makefile for the new entry point and installation path.
