+++ Makefile
@@ -124,6 +124,8 @@ examples_install:
 	${INSTALL} -m 0644 examples/dnsdoctor/*.lua ${SCRIPTS_INSTALL_PATH}/examples/dnsdoctor
 	${MKDIR} ${SCRIPTS_INSTALL_PATH}/examples/systrack
 	${INSTALL} -m 0644 examples/systrack/*.lua ${SCRIPTS_INSTALL_PATH}/examples/systrack
+	${MKDIR} ${SCRIPTS_INSTALL_PATH}/examples/spyglass
+	${INSTALL} -m 0644 examples/spyglass/*.lua ${SCRIPTS_INSTALL_PATH}/examples/spyglass
 	${MKDIR} ${SCRIPTS_INSTALL_PATH}/examples/tcpreject
 	${INSTALL} -m 0644 examples/tcpreject/*.lua ${SCRIPTS_INSTALL_PATH}/examples/tcpreject
